refactor(server): prevent any connected account provider to pass through the imap
